+++ SockPair.hpp
@@ -47,8 +47,8 @@ public:
 	CSockConfigPtr	m_pConfig;		// The socket config.
 	CPath			m_strSendFile;	// Path of log file for data sent.
 	CPath			m_strRecvFile;	// Path of log file for data recieved.
-	uint			m_nBytesSent;	// Total bytes sent.
-	uint			m_nBytesRecv;	// Total bytes recieved.
+	size_t			m_nBytesSent;	// Total bytes sent.
+	size_t			m_nBytesRecv;	// Total bytes recieved.
 
 private:
 	//
